+++ spec/support/stub_config_keys.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module StubConfigKeys
+  extend RSpec::SharedContext
+
+  before do
+    allow(Uploadcare.config).to receive_message_chain(:public_key).and_return('demopublickey')
+    allow(Uploadcare.config).to receive_message_chain(:secret_key).and_return('demosecretkey')
+  end
+end